requirements.txt
openai>=1.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
